pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "RAG Chatbot Backend for Docusaurus AI ROS Book"
authors = [{name = "AI Engineer", email = "engineer@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=4.0.0",
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0",
    "openai>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]